fix: complete AI28 runtime cleanup
